Add files via upload
This file demonstrates the sequential code (data source link is attached in the file) to study the following objective. To assess the association between BMI category (overweight/obese) and the likelihood of being diagnosed with Coronary Heart Disease (CAD). The aim is to determine whether individuals categorized as overweight or obese are at increased risk of CAD compared to those with a healthy BMI. Additionally to determine if there is any significant role of Sex and Income in influencing the outcome.  This task is a part for final project in MPH at Brown School.